+++ planner/src/main/java/edu/planner/repositories/ISubjectRepo.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import edu.planner.models.Subject;
 
-public interface ISubjectRepo extends PagingAndSortingRepository<Subject, Integer> {
+public interface ISubjectRepo extends PagingAndSortingRepository<Subject, Long> {
 
 	public Page<Subject> findByNameContaining(Pageable page, String description);
 
-	public Iterable<Subject> findByCoursesIdIs(Integer coursesId);
+	public Iterable<Subject> findByCoursesIdIs(Long coursesId);
 }
